Fixed an issue where decimals would be truncated.
Removed the code from the calculate() function which turned the output
into an integer, no matter what the output was. Also, I refined the
regex pattern to include the decimal and whatever comes after.

For the future: I can add functionality to calculate exponents, as well
as any other math function. This is a super modular calculator!